auto_trading_system/requirements.txt
yfinance
pandas
numpy
scikit-learn
matplotlib
gspread
oauth2client
python-telegram-bot
python-dotenv
ta
xgboost
